requirements.txt
aiohttp==3.13.2
akqmt==0.1.0
akshare==1.18.2
baostock==0.8.9
beautifulsoup4==4.14.3
booktype==1.5
browser_use==0.11.2
colorlog==6.10.1
curl_cffi==0.14.0
dbutils==3.1.2
jsonpath==0.82.2
langchain==1.2.3
langchain_community==0.4.1
langchain_core==1.2.7
langchain_deepseek==1.0.1
langchain_experimental==0.4.1
langchain_openai==1.1.7
langgraph==1.0.5
lxml==6.0.2
mini_racer==0.13.1
MySQL-python==1.2.5
nest_asyncio==1.6.0
numpy==2.4.1
openai==2.15.0
pandas==2.3.3
Pillow==12.1.0
pydantic==2.12.5
pymongo_amplidata==3.6.0.post1
python-dotenv==1.2.1
PyYAML==6.0.3
Requests==2.32.5
selenium==4.39.0
taos==1.0.1
tenacity==9.1.2
tiktoken==0.12.0
torch==2.9.1+cu130
transformers==4.57.3
typing_extensions==4.15.0
urllib3==2.6.3
